Background and floor
BUG! I know how to use setOrigin to make it scale for the world but the face that u have menuback grounds at this hour of the night I'm too drained to try to understand the difference so just the background images have to change. u can keep the menu with the default blue background dho.